Use the python of the build environment instead of system python
It seems that libArcus still prefers finding the system python (in my case Python3.8.2) instead of the one built in our environment (v3.8.8), even though we specifically link 'our' version of the executable while configuring cura-build. Hence, the same link to the Python executable needs to be made when configuring libArcus.